docs(SafeArea): 调整demo
